% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{browse_gc_credentials}
\alias{browse_gc_credentials}
\title{Visit Google Cloud Credentials page}
\usage{
browse_gc_credentials()
}
\value{
The url of the "Credentials" page, invisibly.
}
\description{
Launch the Google Cloud "Credentials" page (where you can configure OAuth 2.0
Clients). After you configure a client, copy/paste the Client ID and Client
Secret into the \code{YOUTUBE_CLIENT_ID} and \code{YOUTUBE_CLIENT_SECRET} environment
variables. We recommend placing these environment variables in your
\code{.Renviron} file.
}
\examples{
gc_credentials_url <- browse_gc_credentials()
# Copy/paste values from your client.
Sys.setenv(YOUTUBE_CLIENT_ID = "12345-ab12c.apps.googleusercontent.com")
Sys.setenv(YOUTUBE_CLIENT_SECRET = "ABCD-eFg_H")
}
