% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string_contains.R
\name{comments_notcontains_and}
\alias{comments_notcontains_and}
\title{comments_notcontains_and}
\usage{
comments_notcontains_and(vector, dataset.df)
}
\arguments{
\item{vector}{}

\item{dataset.df}{}
}
\value{
dataset11.df
}
\description{
comments_notcontains_and
}

