% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string_contains.R
\name{comments_notcontains_or}
\alias{comments_notcontains_or}
\title{comments_notcontains_or}
\usage{
comments_notcontains_or(vector, dataset.df)
}
\arguments{
\item{vector}{}

\item{dataset.df}{}
}
\value{
dataset00.df
}
\description{
comments_notcontains_or
}

