% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{datetime}
\alias{datetime}
\title{Datetime vectors}
\usage{
datetime()
}
\value{
A length-0 POSIXct object.
}
\description{
Create an object of type "POSIXct". This function is used to provide a
length-0 placeholder value in function definitions.
}
\examples{
datetime()
}
