% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string_contains.R
\name{description_contains_and}
\alias{description_contains_and}
\title{description_contains_and}
\usage{
description_contains_and(vector, dataset.df)
}
\arguments{
\item{vector}{}

\item{dataset.df}{}
}
\value{
dataset.df111
}
\description{
description_contains_and
}

