% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string_contains.R
\name{description_notcontains_and}
\alias{description_notcontains_and}
\title{description_notcontains_and}
\usage{
description_notcontains_and(vector, dataset.df)
}
\arguments{
\item{vector}{}

\item{dataset.df}{}
}
\value{
dataset.df1111
}
\description{
description_notcontains_and
}

