% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string_contains.R
\name{description_notcontains_or}
\alias{description_notcontains_or}
\title{description_notcontains_or}
\usage{
description_notcontains_or(vector, dataset.df)
}
\arguments{
\item{vector}{}

\item{dataset.df}{}
}
\value{
dataset.df0000
}
\description{
description_notcontains_or
}

