% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.add_body}
\alias{.add_body}
\title{Add the body to the request}
\usage{
.add_body(request, body)
}
\arguments{
\item{request}{The rest of the request.}

\item{body}{An optional list of parameters to pass in the body portion of the
request. Should have additional class "json" or "multipart". See the
associated http2 functions for details.}
}
\value{
The request with the body appropriately added.
}
\description{
Add the body to the request
}
\keyword{internal}
