% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.call_youtube_api}
\alias{.call_youtube_api}
\title{Call the YouTube API}
\usage{
.call_youtube_api(
  endpoint,
  query,
  body,
  method,
  token = fetch_token(),
  base_url = c("basic", "upload", "resumable_upload")
)
}
\arguments{
\item{endpoint}{The path to an endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{An optional list of parameters to pass in the query portion of
the request.}

\item{body}{An optional list of parameters to pass in the body portion of the
request. Should have additional class "json" or "multipart". See the
associated http2 functions for details.}

\item{method}{If the method is something other than GET or POST, supply it.
Case is ignored.}

\item{token}{Your YouTube API token. We recommend passing this as a
\code{YOUTUBE_TOKEN} environment variable.}

\item{base_url}{Which family of URLs to use. Almost everything will use the
default basic URL.}
}
\value{
The result of the call.
}
\description{
This is unexported because I intend to use it from endpoint-specific
functions. Only export the resulting functions.
}
\keyword{internal}
