% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{.compact}
\alias{.compact}
\title{Discard empty elements}
\usage{
.compact(lst, depth = 1, max_depth = 20)
}
\arguments{
\item{lst}{A (nested) list to filter.}

\item{depth}{The current recursion depth.}

\item{max_depth}{The maximum recursion depth.}
}
\value{
The list, minus empty elements and branches.
}
\description{
Discard empty elements in nested lists.
}
\keyword{internal}
