% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{.detect_parts}
\alias{.detect_parts}
\title{Stringify the names of a list}
\usage{
.detect_parts(lst)
}
\arguments{
\item{lst}{A named list object. This object will be compacted, and then the
names of any remaining components will be returned, as a comma-separated
list.}
}
\value{
A comma-separated character, such as "this,that".
}
\description{
Stringify the names of a list
}
\keyword{internal}
