% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{.get_token_noninteractive}
\alias{.get_token_noninteractive}
\title{Retrieve a YouTube OAuth token if possible}
\usage{
.get_token_noninteractive(client, refresh_token = NULL)
}
\arguments{
\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{refresh_token}{A refresh token associated with this \code{client}.}
}
\value{
A YouTube OAuth token, or \code{NULL}.
}
\description{
Retrieve a YouTube OAuth token if possible
}
\keyword{internal}
