% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_responses.R
\name{.new_youtube_response}
\alias{.new_youtube_response}
\title{Construct a youtube_response}
\usage{
.new_youtube_response(response_data)
}
\arguments{
\item{response_data}{The \code{data} portion of a response from
\code{httr2::resp_body_json()}.}
}
\value{
An object with additional class "youtube_response", or \code{NULL}.
}
\description{
Construct a youtube_response
}
\keyword{internal}
