% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_responses.R
\name{.parse_response}
\alias{.parse_response}
\title{Parse the returned response}
\usage{
.parse_response(response)
}
\arguments{
\item{response}{A raw response returned from the YouTube api.}
}
\value{
A \code{youtube_response} \code{list} object.
}
\description{
Parse the returned response
}
\keyword{internal}
