% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.prepare_request}
\alias{.prepare_request}
\title{Combine request pieces}
\usage{
.prepare_request(endpoint, query, body, method, token)
}
\arguments{
\item{endpoint}{The path to an endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{An optional list of parameters to pass in the query portion of
the request.}

\item{body}{An optional list of parameters to pass in the body portion of the
request. Should have additional class "json" or "multipart". See the
associated http2 functions for details.}

\item{method}{If the method is something other than GET or POST, supply it.
Case is ignored.}

\item{token}{Your YouTube API token. We recommend passing this as a
\code{YOUTUBE_TOKEN} environment variable.}
}
\value{
A request ready to perform.
}
\description{
Combine request pieces
}
\keyword{internal}
