% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{.yt_req_auth}
\alias{.yt_req_auth}
\title{YouTube OAuth authentication}
\usage{
.yt_req_auth(
  request,
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{request}{A \code{\link[httr2:request]{httr2::request()}}.}

\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \code{vignette("oauth")}}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}
}
\value{
A \code{\link[httr2:request]{httr2::request()}} with oauth authentication information.
}
\description{
This function is the main way authentication is handled in this package. It
tries to find a token non-interactively if possible, but bothers the user if
necessary.
}
\keyword{internal}
