% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{fetch_token}
\alias{fetch_token}
\title{Fetch a YouTube API token}
\usage{
fetch_token(google_client = construct_client(), force = FALSE)
}
\arguments{
\item{google_client}{A client created with \code{\link[=construct_client]{construct_client()}}.}

\item{force}{A logical indicating whether to force a refresh of the token.}
}
\value{
A YouTube token, invisibly.
}
\description{
Load or generate a \code{YOUTUBE_TOKEN} for use in the other functions in this
package.
}
\examples{
token <- fetch_token()
}
