% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_id.R
\name{get_channel_id}
\alias{get_channel_id}
\title{Get Channel Id by Name}
\usage{
get_channel_id(
  handle,
  max_results = 100,
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{handle}{youtube handle (not including the @ symbol)}

\item{max_results}{The maximum number of results to return.}

\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \code{vignette("oauth")}}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}
}
\value{
The channel ID for that username, string.
}
\description{
Retrieves the Youtube channel ID from the username.
}
