% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_comments_sentiment.R
\name{get_comments_sentiment}
\alias{get_comments_sentiment}
\title{get_comments_sentiment}
\usage{
get_comments_sentiment(dataset)
}
\arguments{
\item{dataset}{is a dataframe}
}
\value{
dataset.df is another dataframe
}
\description{
get_comments_sentiment
}

