% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{get_my_channel_detail_playlist_ids}
\alias{get_my_channel_detail_playlist_ids}
\title{Get my channel detail playlist ids}
\usage{
get_my_channel_detail_playlist_ids(
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \url{https://httr2.r-lib.org/articles/oauth.html}.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}
}
\value{
The \emph{contentDetails} object encapsulates information about the
channel's content.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   watchHistory \tab character \tab The ID of the playlist that contains the channel's watch history. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   likes \tab character \tab The ID of the playlist that contains the channel's liked videos. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   favorites \tab character \tab The ID of the playlist that contains the channel's favorite videos. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   watchLater \tab character \tab The ID of the playlist that contains the channel's watch later playlist. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   uploads \tab character \tab The ID of the playlist that contains the channel's uploaded videos. Use the videos.insert method to upload new videos and the videos.delete method to delete previously uploaded videos. \cr
}
}
\description{
Get ids of playlists that contain details about the content of a channel.
}
\examples{
\dontshow{if (yt_has_client_envvars() && interactive()) withAutoprint(\{ # examplesIf}
get_my_channel_details()
\dontshow{\}) # examplesIf}
}
