% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_ids.R
\name{get_playlist_ids}
\alias{get_playlist_ids}
\title{Get Playlist Ids}
\usage{
get_playlist_ids(
  client,
  token_url = "https://oauth2.googleapis.com/token",
  auth_url = "https://accounts.google.com/o/oauth2/v2/auth",
  scope = "https://www.googleapis.com/auth/youtube"
)
}
\arguments{
\item{client}{an httr2 oauth2 client for the google api}

\item{token_url}{google's oauth2 token url}

\item{auth_url}{google's oauth2 auth url}

\item{scope}{token scope, default is youtube}
}
\value{
list of playlist ids for uploads
}
\description{
Retrieves the upload playlist ids.
}
\examples{

google_client <- construct_client()

playlist_ids <- get_playlist_ids(google_client)
}
