% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_items.R
\name{get_playlist_video_ids}
\alias{get_playlist_video_ids}
\title{Get playlist video ids}
\usage{
get_playlist_video_ids(
  playlist_id,
  max_results = 100,
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{playlist_id}{The ID of a playlist.}

\item{max_results}{The maximum number of results to return.}

\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \url{https://httr2.r-lib.org/articles/oauth.html}.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}
}
\value{
A character vector of video ids.
}
\description{
Retrieves video ids for a particular playlist.
}
\examples{
\dontshow{if (yt_has_client_envvars() && interactive()) withAutoprint(\{ # examplesIf}
get_playlist_video_ids(playlist_id = get_upload_playlist_id())
\dontshow{\}) # examplesIf}
}
