% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_videos_for_playlistId.R
\name{get_videos_for_playlistId}
\alias{get_videos_for_playlistId}
\title{Get videos for given playlistId}
\usage{
get_videos_for_playlistId(
  playlistId = NULL,
  max_results = 100,
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{playlistId}{Youtube assigned id for given playlist.}

\item{max_results}{The maximum number of results to return.}

\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \code{vignette("oauth")}}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}
}
\value{
A list of playlists, each with title and description.
}
\description{
Get videos for given playlistId
}
\examples{
\dontshow{if (yt_has_client_envvars() && interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_videos_for_playlistId()

playlistId <<- "PLbcglKxZP5PMU2rNPBpYIzLTgzd5aOHw2"
\dontshow{\}) # examplesIf}
}
