% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getComments.R
\name{getComments}
\alias{getComments}
\title{this is to get the first 1000 comments}
\usage{
getComments(x)
}
\arguments{
\item{x}{is a string vector}
}
\value{
dataset.tdf is a table dataframe
}
\description{
this is to get the first 1000 comments
}

