% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getVideoID.R
\name{getVideoID}
\alias{getVideoID}
\title{This function returns videoid}
\usage{
getVideoID(channelid)
}
\arguments{
\item{channelid}{}
}
\value{
dataset.ls
}
\description{
This function returns videoid
}

