% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{has_youtube_client_envvars}
\alias{has_youtube_client_envvars}
\title{Check for YouTube client environment variables}
\usage{
has_youtube_client_envvars(
  client_id = Sys.getenv("YOUTUBE_CLIENT_ID"),
  client_secret = Sys.getenv("YOUTUBE_CLIENT_SECRET")
)
}
\arguments{
\item{client_id}{A \href{https://console.cloud.google.com/apis/credentials}{Google Cloud OAuth 2.0} client ID. We
recommend you save it as an environment variable, \code{YOUTUBE_CLIENT_ID}.}

\item{client_secret}{A \href{https://console.cloud.google.com/apis/credentials}{Google Cloud OAuth 2.0} client secret. We
recommend you save it as an environment variable, \code{YOUTUBE_CLIENT_SECRET}.}
}
\value{
A logical indicating whether the variables are available.
}
\description{
Reports whether YouTube client environment variables are set (or, if
arguments are provided, check that they have characters).
}
\examples{
has_youtube_client_envvars()
has_youtube_client_envvars("an_id_string", "a_secret_string")
}
