% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/join_SnD.R
\name{join_SnD}
\alias{join_SnD}
\title{description}
\usage{
join_SnD(df1, df2)
}
\arguments{
\item{df1}{is a static dataframe obtained from YouTube}

\item{df2}{is a dataframe of dynamic data obtained from YouTube}
}
\value{
df combined dataframe about a channel
}
\description{
description
}

