% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_videos_for_playlistId.R
\name{make_tibble}
\alias{make_tibble}
\title{Return tibble for videos for given playlist}
\usage{
make_tibble(playlistId = playlistId)
}
\arguments{
\item{playlistId}{Youtube assigned id for given playlist.}
}
\description{
Return tibble for videos for given playlist
}
