% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{This a general retry function with a wrapper with try() and trycatch().
In general cases this functions is to be used with readLines or
rjson::fromJSON.}
\usage{
retry(.FUN, max.attempts = 5, sleep.seconds = 60)
}
\arguments{
\item{.FUN}{is a function which is going to be executed.}
}
\value{
it returns API content
}
\description{
This a general retry function with a wrapper with try() and trycatch().
In general cases this functions is to be used with readLines or
rjson::fromJSON.
}

