% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string_contains.R
\name{title_contains_and}
\alias{title_contains_and}
\title{title_contains_and}
\usage{
title_contains_and(vector, dataset.df)
}
\arguments{
\item{vector}{}

\item{dataset.df}{}
}
\value{
datset.df1
}
\description{
title_contains_and
}

