% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string_contains.R
\name{title_notcontains_and}
\alias{title_notcontains_and}
\title{title_notcontains_and}
\usage{
title_notcontains_and(vector, dataset.df)
}
\arguments{
\item{vector}{}

\item{dataset.df}{}
}
\value{
dataset.df00
}
\description{
title_notcontains_and
}

