% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twittershare.R
\name{twittershare}
\alias{twittershare}
\title{description}
\usage{
twittershare(vidID)
}
\arguments{
\item{vidID}{string variable}
}
\value{
twitter_share numeric veriable
}
\description{
description
}

