% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_video.R
\name{upload_video}
\alias{upload_video}
\title{Upload a video to youtube}
\usage{
upload_video(client, snippet, video_path)
}
\arguments{
\item{token_url}{Google's oauth url}
}
\value{
an httr2 oauth client object
}
\description{
Takes a client and a video file path and uploads the video to youtube. It returns the video id.ß
}
\details{
Requires the following environment variables: YOUTUBE_CLIENT_ID, YOUTUBE_CLIENT_SECRET
}
\examples{

google_client <- create_client()


}
