% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeR-package.R
\docType{package}
\name{youtubeR-package}
\alias{youtubeR}
\alias{youtubeR-package}
\title{youtubeR: An Interface to the YouTube API}
\description{
Manage YouTube videos from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kevin-m-kent.github.io/youtubeR/}
  \item \url{https://github.com/kevin-m-kent/youtubeR}
  \item Report bugs at \url{https://github.com/kevin-m-kent/youtubeR/issues}
}

}
\author{
\strong{Maintainer}: Kevin Kent \email{kentmkevin@gmail.com}

Authors:
\itemize{
  \item Kevin Kent \email{kentmkevin@gmail.com}
  \item Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
}

}
\keyword{internal}
