% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{yt_call_api}
\alias{yt_call_api}
\title{Call the YouTube API}
\usage{
yt_call_api(
  endpoint,
  query = NULL,
  body = NULL,
  method = NULL,
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL,
  base_url = c("basic", "upload", "resumable_upload")
)
}
\arguments{
\item{endpoint}{The path to an endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{An optional list of parameters to pass in the query portion of
the request.}

\item{body}{An optional list of parameters to pass in the body portion of the
request.}

\item{method}{If the method is something other than GET or POST, supply it.
Case is ignored.}

\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \code{vignette("oauth")}}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}

\item{base_url}{Which family of URLs to use. Almost everything will use the
default basic URL.}
}
\value{
The result of the call.
}
\description{
If you use this function, please \href{https://github.com/kevin-m-kent/youtubeR/issues}{open an issue in the youtubeR GitHub repository} describing your
use case. We intend to wrap all endpoints such that this function will not be
necessary for users.
}
