% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_channels.R
\name{yt_channels}
\alias{yt_channels}
\title{YouTube Channels API node}
\usage{
yt_channels(part, id, maxResults = NULL, pageToken = NULL, api.key)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more channel resource properties that the API response will include. Valid parameters can be found here: https://developers.google.com/youtube/v3/docs/channels/list.}

\item{id}{The id property specifies the channel's YouTube channel ID.}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set. Acceptable values are 0 to 50, inclusive. The default value is 5.}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned. In an API response, the nextPageToken and prevPageToken properties identify other pages that could be retrieved.}

\item{api.key}{The api.key parameter species the key used to access the API.}
}
\value{
Returns an object with the API call results.
}
\description{
Access the YouTube Channels API node
}
\examples{
\dontrun{
 library(youTubeR)

 test <- yt_channels(part = c("id", "snippet", "localizations", "statistics", "brandingSettings", "topicDetails"),
                          id = "UCJy_cQ9QwYihJ4kIQpvdEAA",
                          maxResults = 50,
                          api.key = INSERT_API_KEY_HERE)

}
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
}
