% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_comment_threads.R
\name{yt_comment_threads}
\alias{yt_comment_threads}
\title{YouTube Comment Threads API node}
\usage{
yt_comment_threads(
  videoId,
  order = NULL,
  searchTerms = NULL,
  maxResults = NULL,
  pageToken = NULL,
  textFormat = NULL,
  api.key
)
}
\arguments{
\item{videoId}{The videoId parameter instructs the API to return comment threads associated with the specified video ID}

\item{order}{The order parameter specifies the order in which the API response should list comment threads. Valid values are "time" and "relevance".}

\item{searchTerms}{The searchTerms parameter instructs the API to limit the API response to only contain comments that contain the specified search terms.}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set. Acceptable values are 1 to 100 inclusive.}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned. In an API response, the nextPageToken property identifies the next page of the result that can be retrieved.}

\item{textFormat}{Set this parameter's value to html or plainText to instruct the API to return the comments left by users in html formatted or in plain text.}

\item{api.key}{The api.key parameter species the key used to access the API.}
}
\value{
Returns an object with the API call results.
}
\description{
Access the YouTube Comment Threads API node
}
\examples{
\dontrun{
comments <- yt_comment_threads(videoId = INSETR_VIDEO_ID,
                               maxResults = 100,
                               order = "time",
                               api.key = INSERT_API_KEY)
}
}
\seealso{
\code{\link[utils]{URLencode}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
}
