% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{yt_construct_client}
\alias{yt_construct_client}
\title{Construct a YouTube OAuth client}
\usage{
yt_construct_client(
  client_id = Sys.getenv("YOUTUBE_CLIENT_ID"),
  client_secret = Sys.getenv("YOUTUBE_CLIENT_SECRET")
)
}
\arguments{
\item{client_id}{A \href{https://console.cloud.google.com/apis/credentials}{Google Cloud OAuth 2.0} client ID. We
recommend you save it as an environment variable, \code{YOUTUBE_CLIENT_ID}.}

\item{client_secret}{A \href{https://console.cloud.google.com/apis/credentials}{Google Cloud OAuth 2.0} client secret. We
recommend you save it as an environment variable, \code{YOUTUBE_CLIENT_SECRET}.}
}
\value{
An \code{\link[httr2:oauth_client]{httr2::oauth_client()}} object.
}
\description{
Builds the OAuth client object for google apis.
}
\examples{
client <- yt_construct_client()
}
