% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_schemas.R
\name{yt_schema_video_status}
\alias{yt_schema_video_status}
\title{Video status}
\usage{
yt_schema_video_status(
  embeddable = logical(),
  license = character(),
  privacy_status = character(),
  public_stats_viewable = logical(),
  publish_at = datetime(),
  self_declared_made_for_kids = logical()
)
}
\arguments{
\item{embeddable}{(logical) This value indicates if the video can be embedded
on another website.}

\item{license}{(factor scalar) The video's license. values: youtube,
creativeCommon}

\item{privacy_status}{(factor scalar) The video's privacy status. values:
public, unlisted, private}

\item{public_stats_viewable}{(logical) This value indicates if the extended
video statistics on the watch page can be viewed by everyone. Note that the
view count, likes, etc will still be visible if this is disabled.}

\item{publish_at}{(datetime scalar) The date and time when the video is
scheduled to publish. It can be set only if the privacy status of the video
is private.}

\item{self_declared_made_for_kids}{(logical) Whether the video is made for
kids.}
}
\value{
A list with any non-zero-length properties.
}
\description{
The status object contains information about the video's uploading,
processing, and privacy statuses.
}
\examples{
yt_schema_video_status()
yt_schema_video_status(embeddable = TRUE)
}
