% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_search.R
\name{yt_search}
\alias{yt_search}
\title{YouTube Search API node}
\usage{
yt_search(
  query,
  type = "video",
  location = NULL,
  locationRadius = NULL,
  radiusUnit = NULL,
  maxResults = NULL,
  pageToken = NULL,
  publishedAfter = NULL,
  publishedBefore = NULL,
  api.key
)
}
\arguments{
\item{query}{The query parameter specifies the query term to search for.}

\item{type}{The type parameter restricts a search query to only retrieve a particular type of resource. The value is a comma-separated list of resource types. The default value is "video", but also recognizes "channel" and "playlist".}

\item{location}{The location parameter, in conjunction with the locationRadius and radiusUnit parameters, defines a circular geographic area and also restricts a search to videos that specify, in their metadata, a geographic location that falls within that area. The parameter value is a string that specifies latitude/longitude coordinates e.g. (37.42307,-122.08427).}

\item{locationRadius}{The locationRadius parameter, in conjunction with the location parameter, defines a circular geographic area. The API does not support locationRadius parameter values larger than 1000 kilometers.}

\item{radiusUnit}{Measurement unit for the locationRadius parameter. Valid measurement units are m, km, ft, and mi.}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set. Acceptable values are 0 to 50, inclusive.}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned. In an API response, the nextPageToken and prevPageToken properties identify other pages that could be retrieved.}

\item{publishedAfter}{The publishedAfter parameter indicates that the API response should only contain resources created at or after the specified time. Values should be provided in POSIXct or Date format.}

\item{publishedBefore}{The publishedBefore parameter indicates that the API response should only contain resources created before or at the specified time. Values should be provided in POSIXct or Date format.}

\item{api.key}{The api.key parameter species the key used to access the API.}
}
\value{
Returns an object with the API call results.
}
\description{
Access the YouTube Search API node
}
\examples{
\dontrun{
video_search <- yt_search_list(query = INSERT_QUERY,
                        maxResults = 50,
                        publishedAfter = as.Date("2021/10/01"),
                        api.key = INSERT_API_KEY)
}
}
\seealso{
\code{\link[utils]{URLencode}}
 \code{\link[lubridate]{is.POSIXt}},\code{\link[lubridate]{is.Date}}
 \code{\link[anytime]{iso8601}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
}
