% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_schemas.R
\name{yt_video_localization}
\alias{yt_video_localization}
\title{Video localization}
\usage{
yt_video_localization(description = character(), title = character())
}
\arguments{
\item{description}{(character scalar) Localized version of the video's
description.}

\item{title}{(character scalar) Localized version of the video's title.}
}
\value{
A list with any non-zero-length properties.
}
\description{
Localized versions of certain video properties (e.g. title).
}
\examples{
yt_video_localization()
yt_video_localization(description = "A description.")
}
