% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_schemas.R
\name{yt_video_snippet}
\alias{yt_video_snippet}
\title{Video snippet}
\usage{
yt_video_snippet(
  title = character(),
  description = character(),
  tags = character(),
  category_id = character(),
  default_language = character()
)
}
\arguments{
\item{title}{(character scalar) The video's title.}

\item{description}{(character scalar) The video's description.}

\item{tags}{(character vector) A list of keyword tags associated with the
video. Tags may contain spaces.}

\item{category_id}{(character scalar) The YouTube video category associated
with the video.}

\item{default_language}{(character scalar) The language of the videos's
default snippet.}
}
\value{
A list with any non-zero-length properties.
}
\description{
Basic details about a video, including title, description, uploader,
thumbnails and category.
}
\examples{
yt_video_snippet()
yt_video_snippet(title = "An example video", tags = c("a", "b"))
}
