% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{yt_videos_insert}
\alias{yt_videos_insert}
\title{Insert video}
\usage{
yt_videos_insert(
  video_path,
  snippet = yt_video_snippet(),
  localizations = list(),
  status = yt_video_status(),
  recording_date = datetime(),
  token = fetch_token()
)
}
\arguments{
\item{video_path}{(character scalar) Path to a video file to upload.}

\item{snippet}{Basic details about a video, including title, description,
uploader, thumbnails and category. See \code{\link[=yt_video_snippet]{yt_video_snippet()}} for details.}

\item{localizations}{A list of named localizations objects. The localizations
object contains localized versions of the basic details about the video,
such as its title and description. The name should be a valid language key,
which is a \href{http://www.rfc-editor.org/rfc/bcp/bcp47.txt}{BCP-47} language
code. See \code{\link[=yt_video_localization]{yt_video_localization()}} for details on the format for each
localization.}

\item{status}{Basic details about a video category, such as its localized
title. See \code{\link[=yt_video_status]{yt_video_status()}} for details.}

\item{recording_date}{(datetime scalar) The date and time when the video was
recorded.}

\item{token}{Your YouTube API token. We recommend passing this as a
\code{YOUTUBE_TOKEN} environment variable.}
}
\value{
The id of the uploaded video.
}
\description{
Inserts a new resource into this collection.
}
