% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.prepare_body}
\alias{.prepare_body}
\title{Prepare the body of a call}
\usage{
.prepare_body(body, type = c("json", "multipart"), mime_type)
}
\arguments{
\item{body}{An object to use as the body of the request.}

\item{type}{Whether the request is "json" or "multipart".}

\item{mime_type}{The mime_type of any file included in the body.}
}
\value{
A prepared body list object with a "json" or "multipart" subclass.
}
\description{
Prepare the body of a call
}
\keyword{internal}
