% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{get_my_channel_detail_playlist_ids}
\alias{get_my_channel_detail_playlist_ids}
\title{Get my channel detail playlist ids}
\usage{
get_my_channel_detail_playlist_ids(token = fetch_token())
}
\arguments{
\item{token}{Your YouTube API token. We recommend passing this as a
\code{YOUTUBE_TOKEN} environment variable.}
}
\value{
The \emph{contentDetails} object encapsulates information about the
channel's content.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   watchHistory \tab character \tab The ID of the playlist that contains the channel's watch history. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   likes \tab character \tab The ID of the playlist that contains the channel's liked videos. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   favorites \tab character \tab The ID of the playlist that contains the channel's favorite videos. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   watchLater \tab character \tab The ID of the playlist that contains the channel's watch later playlist. Use the playlistItems.insert and playlistItems.delete to add or remove items from that list. \cr
   uploads \tab character \tab The ID of the playlist that contains the channel's uploaded videos. Use the videos.insert method to upload new videos and the videos.delete method to delete previously uploaded videos. \cr
}
}
\description{
Get ids of playlists that contain details about the content of a channel.
}
\examples{
\dontshow{if (has_youtube_client_envvars() && interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_my_channel_details()
\dontshow{\}) # examplesIf}
}
