% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_items.R
\name{get_playlist_video_ids}
\alias{get_playlist_video_ids}
\title{Get playlist video ids}
\usage{
get_playlist_video_ids(playlist_id, max_results = 100, token = fetch_token())
}
\arguments{
\item{playlist_id}{The ID of a playlist.}

\item{max_results}{The maximum number of results to return.}

\item{token}{Your YouTube API token. We recommend passing this as a
\code{YOUTUBE_TOKEN} environment variable.}
}
\value{
A character vector of video ids.
}
\description{
Retrieves video ids for a particular playlist.
}
\examples{
\dontshow{if (has_youtube_client_envvars() && interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_playlist_video_ids(playlist_id = get_upload_playlist_id())
\dontshow{\}) # examplesIf}
}
