% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{get_upload_playlist_id}
\alias{get_upload_playlist_id}
\title{Get Playlist Ids}
\usage{
get_upload_playlist_id(token = fetch_token())
}
\arguments{
\item{token}{Your YouTube API token. We recommend passing this as a
\code{YOUTUBE_TOKEN} environment variable.}
}
\value{
list of playlist ids for uploads
}
\description{
Retrieves the upload playlist ids.
}
\examples{
\dontshow{if (has_youtube_client_envvars() && interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_upload_playlist_id()
\dontshow{\}) # examplesIf}
}
