% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{get_video_processing_details}
\alias{get_video_processing_details}
\title{Get video processing details}
\usage{
get_video_processing_details(video_ids, token = fetch_token())
}
\arguments{
\item{video_ids}{A character vector of video ids.}

\item{token}{Your YouTube API token. We recommend passing this as a
\code{YOUTUBE_TOKEN} environment variable.}
}
\value{
A list of processing status details. processingStatus is probably of
most interest
}
\description{
Retrieves the processing details for a vector of video IDs
}
\examples{
\dontshow{if (has_youtube_client_envvars() && interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_upload_playlist_id() |>
  get_playlist_items() |>
  get_video_processing_details()
\dontshow{\}) # examplesIf}
}
