% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{retry
This is a general error handeling function. base function used in this function
is try() and trycatch()
In package this function will mainly handel erros of the following functions
httr::GET, base::readLines or rjson::fromJSON}
\usage{
retry(.FUN, url, max.attempts = 5, sleep.seconds = 30)
}
\arguments{
\item{.FUN}{is a string variable. It is the string form of supplied function which is going to be
evaluated and the function retry will deal with the errors commited in this process.}

\item{url}{is a string variable. This is going to be used as a input parameter in the .FUN function
supplied by the user.}
}
\value{
x is a string variable. Mainly x will contain data obtained from api either in XML ,JSON or
simply in text fromat.
}
\description{
retry
This is a general error handeling function. base function used in this function
is try() and trycatch()
In package this function will mainly handel erros of the following functions
httr::GET, base::readLines or rjson::fromJSON
}

