% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_video.R
\name{upload_video}
\alias{upload_video}
\title{Upload a video to youtube}
\usage{
upload_video(
  client,
  snippet,
  video_path,
  scope = "https://www.googleapis.com/auth/youtube",
  token_url = "https://oauth2.googleapis.com/token",
  auth_url = "https://accounts.google.com/o/oauth2/v2/auth"
)
}
\arguments{
\item{token_url}{Google's oauth url}
}
\value{
an httr2 oauth client object
}
\description{
Takes a client and a video file path and uploads the video to youtube. It returns the video id.ß
}
\details{
Requires the following environment variables: YOUTUBE_CLIENT_ID, YOUTUBE_CLIENT_SECRET
}
\examples{

google_client <- create_client()

snippet <- list(snippet = list("title" = unbox("video kids test"),
                       "description" = unbox("description_test"),
                       "tags" = "kevin,kent"),
 status = list("privacyStatus" = unbox("private"),
              "selfDeclaredMadeForKids" = unbox("false"))) 

video_path <- "my_video.mp4"

upload_video(google_client, snippet, video_path)
}
