% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{yt_authenticate}
\alias{yt_authenticate}
\title{Authenticate with a YouTube OAuth client}
\usage{
yt_authenticate(
  client = yt_construct_client(),
  force = FALSE,
  refresh_token = NULL
)
}
\arguments{
\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{force}{A logical indicating whether to force a refresh of the token.}

\item{refresh_token}{A refresh token associated with this \code{client}. This
parameter exists primarily for testing. If you wish to provide a refresh
token (for example, for automated processes), we recommend setting a
\code{YOUTUBE_REFRESH_TOKEN} environment variable.}
}
\value{
A YouTube OAuth token, invisibly.
}
\description{
Load or generate a YouTube OAuth token for use in the other functions in
this package. The primary use of this function is to cache values early in a
script (so the user can walk away). Otherwise the other functions in this
package will prompt for authentication when needed. Once the values are
cached, the rest of this package will use them by default for that client.
}
\examples{
\dontshow{if (yt_has_client_envvars() && interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
token <- yt_authenticate()
\dontshow{\}) # examplesIf}
}
