% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{yt_videos_delete}
\alias{yt_videos_delete}
\title{Delete video}
\usage{
yt_videos_delete(
  video_id,
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{video_id}{(character scalar) The id parameter specifies the YouTube
video ID for the resource that is being deleted. In a video resource, the
id property specifies the video's ID.}

\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \code{vignette("oauth")}}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}
}
\value{
The id of the deleted video.
}
\description{
Deletes a resource.
}
