% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{yt_videos_insert}
\alias{yt_videos_insert}
\title{Insert video}
\usage{
yt_videos_insert(
  video_path,
  snippet = yt_schema_video_snippet(),
  localizations = list(),
  status = yt_schema_video_status(),
  recording_date = datetime(),
  client = yt_construct_client(),
  cache_disk = getOption("yt_cache_disk", FALSE),
  cache_key = getOption("yt_cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{video_path}{(character scalar) Path to a video file to upload.}

\item{snippet}{Basic details about a video, including title, description,
uploader, thumbnails and category. See \code{\link[=yt_schema_video_snippet]{yt_schema_video_snippet()}} for
details.}

\item{localizations}{A list of named localizations objects. The localizations
object contains localized versions of the basic details about the video,
such as its title and description. The name should be a valid language key,
which is a \href{http://www.rfc-editor.org/rfc/bcp/bcp47.txt}{BCP-47} language
code. See \code{\link[=yt_schema_video_localization]{yt_schema_video_localization()}} for details on the format for
each localization.}

\item{status}{Basic details about a video category, such as its localized
title. See \code{\link[=yt_schema_video_status]{yt_schema_video_status()}} for details.}

\item{recording_date}{(datetime scalar) The date and time when the video was
recorded.}

\item{client}{A YouTube OAuth client created with \code{\link[=yt_construct_client]{yt_construct_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \code{vignette("oauth")}}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A YouTube API OAuth token, or the \code{access_token} string from
such a token. We recommend that you instead supply a \code{client}, in which
case an appropriate token will be located if possible.}
}
\value{
The id of the uploaded video.
}
\description{
Inserts a new resource into this collection.
}
